refactor: reorg of secret resolver
