+++ src/sap_cloud_sdk/core/secret_resolver/constants.py
@@ -3,3 +3,5 @@ Constants for secret resolver module.
 """
 
 BASE_MOUNT_PATH = "/etc/secrets/appfnd"
+BASE_VAR_NAME = "CLOUD_SDK_CFG"
+SERVICE_BINDING_ROOT = "SERVICE_BINDING_ROOT"
